pyproject.toml
[project]
name = "hackathon-zero"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth==2.23.4",
    "google-auth-oauthlib==1.1.0",
    "google-auth-httplib2==0.1.1",
    "google-api-python-client==2.108.0",
    "watchdog==4.0.1",
    "aiohttp>=3.8.0",
    "requests>=2.31.0",
]
